Clarify where BitWarden License is installed in Mac Keychain
When BitWarden Developers are installing Licensing for local & self hosted instances, Mac Keychain allows the license and key to be installed under different groups and defaults to "Local Items". The current guide does not make it immediately obvious that the certifications should be stored under "Default Keychains > login" in order for it to be recognized during "Build and Run the Server" steps.

This change adds a note under "Install Licensing Certificate" steps that explicitly states where the licensing files should be saved.

Fix lint warnings