Clarify where BitWarden License is installed in Mac Keychain (#308)
## Objective

When BitWarden Developers are installing Licensing for local & self hosted instances, Mac Keychain allows the license and key to be installed under different groups and defaults to "Local Items". The current guide does not make it immediately obvious that they must be stored under "Default Keychains > login" in order for it to be recognized during "Build and Run the Server" steps.

This change adds a note in step "Install Licensing Certificate" that explicitly states where the licensing files should be saved.